# api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    MessageErr:
        properties:
            Error:
                type: string
            Message:
                type: string
            Status:
                format: int64
                type: integer
        title: MessageErr represents a error message.
        type: object
        x-go-package: github.com/wexinc/ps-tag-onboarding-go/internal/utils
    User:
        properties:
            age:
                format: int64
                type: integer
                x-go-name: Age
            email:
                type: string
                x-go-name: Email
            first_name:
                type: string
                x-go-name: FirstName
            id:
                format: int64
                type: integer
                x-go-name: Id
            last_name:
                type: string
                x-go-name: LastName
        title: User represents a user.
        type: object
        x-go-package: github.com/wexinc/ps-tag-onboarding-go/internal/model
host: localhost:8089
info:
    title: Tag Onboarding API server.
    version: 1.0.0
paths:
    /users:
        post:
            consumes:
                - application/json
            operationId: saveUser
            parameters:
                - in: body
                  name: User
                  schema:
                    $ref: '#/definitions/User'
            produces:
                - application/json
            responses:
                "201":
                    description: User
                    schema:
                        $ref: '#/definitions/User'
                "400":
                    description: MessageErr
                    schema:
                        $ref: '#/definitions/MessageErr'
                "500":
                    description: MessageErr
                    schema:
                        $ref: '#/definitions/MessageErr'
    /users/:
        get:
            consumes:
                - application/json
            operationId: getAllUser
            produces:
                - application/json
            responses:
                "201":
                    description: User
                    schema:
                        $ref: '#/definitions/User'
                "400":
                    description: MessageErr
                    schema:
                        $ref: '#/definitions/MessageErr'
                "500":
                    description: MessageErr
                    schema:
                        $ref: '#/definitions/MessageErr'
    /users/{userId}:
        delete:
            consumes:
                - application/json
            operationId: deleteUser
            parameters:
                - in: path
                  name: userId
                  required: true
                  type: string
                  x-go-name: UserId
            produces:
                - application/json
            responses:
                "200":
                    description: User
                    schema:
                        $ref: '#/definitions/User'
                "400":
                    description: MessageErr
                    schema:
                        $ref: '#/definitions/MessageErr'
                "404":
                    description: MessageErr
                    schema:
                        $ref: '#/definitions/MessageErr'
                "500":
                    description: MessageErr
                    schema:
                        $ref: '#/definitions/MessageErr'
        get:
            consumes:
                - application/json
            operationId: getUser
            parameters:
                - in: path
                  name: userId
                  required: true
                  type: string
                  x-go-name: UserId
            produces:
                - application/json
            responses:
                "200":
                    description: User
                    schema:
                        $ref: '#/definitions/User'
                "400":
                    description: MessageErr
                    schema:
                        $ref: '#/definitions/MessageErr'
                "404":
                    description: MessageErr
                    schema:
                        $ref: '#/definitions/MessageErr'
                "500":
                    description: MessageErr
                    schema:
                        $ref: '#/definitions/MessageErr'
        put:
            consumes:
                - application/json
            operationId: updateUser
            parameters:
                - in: path
                  name: userId
                  required: true
                  type: string
                  x-go-name: UserId
                - in: body
                  name: User
                  schema:
                    $ref: '#/definitions/User'
            produces:
                - application/json
            responses:
                "200":
                    description: User
                    schema:
                        $ref: '#/definitions/User'
                "400":
                    description: MessageErr
                    schema:
                        $ref: '#/definitions/MessageErr'
                "500":
                    description: MessageErr
                    schema:
                        $ref: '#/definitions/MessageErr'
produces:
    - application/json
schemes:
    - http
swagger: "2.0"